internal/relpath: rename package relfs
