+++ internal/relpath/relpath.go
@@ -1,9 +1,25 @@
-package relfs
+package relpath
 
 import (
 	"os"
 	"path/filepath"
 )
+
+// Abs returns a best-effort absolute representation of path.
+// The returned path may be relative on platforms such as WebAssembly.
+// If the path is not absolute it will be joined with the current
+// working directory to turn it into an absolute path.
+// See [filepath.Abs] for more information.
+func Abs(path string) (string, error) {
+	if filepath.IsAbs(path) {
+		return filepath.Clean(path), nil
+	}
+	wd, err := Getwd()
+	if err != nil {
+		return "", err
+	}
+	return filepath.Join(wd, path), nil
+}
 
 // Getwd returns best-effort path to the current working directory, even on WebAssembly.
 // The path may be absolute, or it may be relative, prefixed with one or more ".." segments.
